CyberWatchAPI/requirements.txt
fastapi[all]~=0.75.1
requests~=2.27.1
python-dotenv~=0.20.0
